File for the code
I usually don't compute it all together because of the great amount of data i have to manage. What i've usually done is to fraction 1) till the ALL variable stored; 2) Till the DX and SDX variable stored; 3) Till the So variable stored. The procedium is simple: save the variables as numpy objects and then call them in the next code.